Solution to Calculate Total and Average Marks of Students

+++ Calculate Total and Average Marks of Students@@67877d61dc107486ad5c2746/code.c
@@ -24,7 +24,7 @@ int main() {
         scanf("%d %49s %f", &students[i].roll_number, students[i].name, &students[i].marks);
     }
 
-    int total_marks;
+    float total_marks;
     float avg_marks;
     cal(students, n, &total_marks, &avg_marks);
